Update root command and allow configuration files to be passed as a flag
