+++ .github/workflows/go.yml
@@ -18,11 +18,6 @@ jobs:
       uses: actions/checkout@v1
     - name: Test
       run: go test -v ./...
-    - name: Set configuration file
-      run: |
-        echo $HOME
-        mkdir -p $HOME/.config/ghorg
-        cp sample-conf.yaml $HOME/.config/ghorg/conf.yaml
     - name: Get dependencies
       run: |
         go get -v -t -d ./...
