+++ cmd/ls.go
@@ -11,10 +11,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func init() {
-	rootCmd.AddCommand(lsCmd)
-}
-
 var lsCmd = &cobra.Command{
 	Use:   "ls [dir]",
 	Short: "List contents of your ghorg home or ghorg directories",
@@ -23,17 +19,6 @@ var lsCmd = &cobra.Command{
 }
 
 func lsFunc(cmd *cobra.Command, argz []string) {
-
-	if cmd.Flags().Changed("color") {
-		colorToggle := cmd.Flag("color").Value.String()
-		if colorToggle == "on" {
-			os.Setenv("GHORG_COLOR", colorToggle)
-		} else {
-			os.Setenv("GHORG_COLOR", "off")
-		}
-
-	}
-
 	if len(argz) == 0 {
 		listGhorgHome()
 	}
@@ -47,15 +32,15 @@ func lsFunc(cmd *cobra.Command, argz []string) {
 }
 
 func listGhorgHome() {
-	ghorgDir := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO")
-	files, err := ioutil.ReadDir(ghorgDir)
+	path := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO")
+	files, err := ioutil.ReadDir(path)
 	if err != nil {
 		colorlog.PrintError("No clones found. Please clone some and try again.")
 	}
 
 	for _, f := range files {
 		if f.IsDir() {
-			colorlog.PrintInfo(ghorgDir + f.Name())
+			colorlog.PrintInfo(path + f.Name())
 		}
 	}
 }
@@ -64,16 +49,16 @@ func listGhorgDir(arg string) {
 
 	arg = strings.ReplaceAll(arg, "-", "_")
 
-	ghorgDir := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + arg
+	path := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + arg
 
-	files, err := ioutil.ReadDir(ghorgDir)
+	files, err := ioutil.ReadDir(path)
 	if err != nil {
 		colorlog.PrintError("No clone found with that name. Please check spelling or reclone.")
 	}
 
 	for _, f := range files {
 		if f.IsDir() {
-			str := filepath.Join(ghorgDir, configs.GetCorrectFilePathSeparator(), f.Name())
+			str := filepath.Join(path, configs.GetCorrectFilePathSeparator(), f.Name())
 			colorlog.PrintSubtleInfo(str)
 		}
 	}
